src/ultra_low_bitrate_codec/requirements.txt
torch
torchaudio
transformers
numpy
scipy
pyyaml
tqdm
soundfile
vector-quantize-pytorch
einops
protobuf
